+++ dnscrypt-proxy/privilage_free.go
@@ -1,0 +1,56 @@
+// +build !windows
+
+package main
+
+import (
+	"os"
+	"os/exec"
+	"os/user"
+	"path/filepath"
+	"strconv"
+	"syscall"
+	"github.com/jedisct1/dlog"
+)
+
+func (proxy *Proxy) dropPrivilege(userStr string, fds []*os.File) {
+	user, err := user.Lookup(userStr)
+	args := os.Args
+
+	if err != nil {
+		dlog.Fatal(err)
+	}
+	uid, err := strconv.Atoi(user.Uid)
+	if err != nil {
+		dlog.Fatal(err)
+	}
+	gid, err := strconv.Atoi(user.Gid)
+	if err != nil {
+		dlog.Fatal(err)
+	}
+	exec_path, err := exec.LookPath(args[0])
+	if err != nil {
+		dlog.Fatal(err)
+	}
+	path, err := filepath.Abs(exec_path)
+	if err != nil {
+		dlog.Fatal(err)
+	}
+
+	// remove arg[0]
+	copy(args[0:], args[0+1:])
+	args[len(args)-1] = ""
+	args = args[:len(args)-1]
+	args = append(args, "-start-child")
+
+	cmd := exec.Command(path, args...)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	cmd.ExtraFiles = fds
+	cmd.SysProcAttr = &syscall.SysProcAttr{}
+	cmd.SysProcAttr.Credential = &syscall.Credential{Uid: uint32(uid), Gid: uint32(gid)}
+	dlog.Notice("Dropping privileges")
+	if err := cmd.Start(); err != nil {
+		dlog.Fatal(err)
+	}
+	os.Exit(0)
+}
