+++ dnscrypt-proxy/privilage_windows.go
@@ -1,0 +1,5 @@
+package main
+
+import "os"
+
+func (proxy *Proxy) dropPrivilege(userStr string, fds []*os.File) {}
